Review response: Mixin moved to presenter
